rename model to dto
