+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/order/service/IOrderService.java
@@ -1,7 +1,7 @@
 package ca.uwaterloo.drinkmasterapi.feature.order.service;
 
-import ca.uwaterloo.drinkmasterapi.feature.order.model.OrderRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.order.model.OrderResponseDTO;
+import ca.uwaterloo.drinkmasterapi.feature.order.dto.OrderRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.order.dto.OrderResponseDTO;
 
 import java.util.List;
