+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/order/service/OrderServiceImpl.java
@@ -5,8 +5,8 @@ import ca.uwaterloo.drinkmasterapi.dao.Drink;
 import ca.uwaterloo.drinkmasterapi.repository.DrinkRepository;
 import ca.uwaterloo.drinkmasterapi.repository.MachineRepository;
 import ca.uwaterloo.drinkmasterapi.dao.Order;
-import ca.uwaterloo.drinkmasterapi.feature.order.model.OrderRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.order.model.OrderResponseDTO;
+import ca.uwaterloo.drinkmasterapi.feature.order.dto.OrderRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.order.dto.OrderResponseDTO;
 import ca.uwaterloo.drinkmasterapi.common.OrderStatusEnum;
 import ca.uwaterloo.drinkmasterapi.repository.OrderRepository;
 import ca.uwaterloo.drinkmasterapi.repository.UserRepository;
