+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/user/service/IUserLoginService.java
@@ -1,8 +1,8 @@
 package ca.uwaterloo.drinkmasterapi.feature.user.service;
 
-import ca.uwaterloo.drinkmasterapi.feature.user.model.LoginRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.UserResponseDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.SignupRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.LoginRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.UserResponseDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.SignupRequestDTO;
 
 public interface IUserLoginService {
     UserResponseDTO signup(SignupRequestDTO signupRequest);
