+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/user/service/UserLoginServiceImpl.java
@@ -1,9 +1,9 @@
 package ca.uwaterloo.drinkmasterapi.feature.user.service;
 
 import ca.uwaterloo.drinkmasterapi.common.InvalidCredentialsException;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.LoginRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.UserResponseDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.SignupRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.LoginRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.UserResponseDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.SignupRequestDTO;
 import ca.uwaterloo.drinkmasterapi.dao.User;
 import ca.uwaterloo.drinkmasterapi.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
