+++ src/test/java/ca/uwaterloo/drinkmasterapi/feature/user/UserLoginServiceImplTest.java
@@ -1,10 +1,10 @@
 package ca.uwaterloo.drinkmasterapi.feature.user;
 
 import ca.uwaterloo.drinkmasterapi.common.InvalidCredentialsException;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.LoginRequestDTO;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.SignupRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.LoginRequestDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.SignupRequestDTO;
 import ca.uwaterloo.drinkmasterapi.dao.User;
-import ca.uwaterloo.drinkmasterapi.feature.user.model.UserResponseDTO;
+import ca.uwaterloo.drinkmasterapi.feature.user.dto.UserResponseDTO;
 import ca.uwaterloo.drinkmasterapi.repository.UserRepository;
 import ca.uwaterloo.drinkmasterapi.feature.user.service.UserLoginServiceImpl;
 import org.junit.jupiter.api.Test;
